requirements.txt
pymysql==0.9.3
lxml==4.4.0
beautifulsoup4==4.7.1
requests==2.21.0
python-dateutil=2.8.0
Django==2.2.4